feat: signup page is updated
