+++ frontend/src/pages/SignUp/index.tsx
@@ -1,203 +1,43 @@
-import {
-	Button,
-	Input,
-	notification,
-	Typography,
-	Switch,
-	Space,
-	Card,
-} from 'antd';
-import signup from 'api/user/signup';
-import ROUTES from 'constants/routes';
-import history from 'lib/history';
-import React, { useState } from 'react';
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
-import { ThunkDispatch } from 'redux-thunk';
-import { UserLoggedIn } from 'store/actions';
-import AppActions from 'types/actions';
-const { Title } = Typography;
+import useFetch from 'hooks/useFetch';
+import React from 'react';
+import SignUpComponent from './SignUp';
+import getVersion from 'api/user/getVersion';
+import { PayloadProps as VersionPayload } from 'types/api/user/getVersion';
+import { PayloadProps as UserPrefPayload } from 'types/api/user/getUserPreference';
 
-import {
-	ButtonContainer,
-	Container,
-	FormWrapper,
-	Label,
-	LeftContainer,
-	Logo,
-	MarginTop,
-} from './styles';
+import Spinner from 'components/Spinner';
+import { Typography } from 'antd';
+import getPreference from 'api/user/getPreference';
 
-const Signup = ({ loggedIn }: SignupProps): JSX.Element => {
-	const [notificationsInstance, Element] = notification.useNotification();
+const SignUp = () => {
+	const versionResponse = useFetch<VersionPayload, undefined>(getVersion);
 
-	const [loading, setLoading] = useState(false);
+	const userPrefResponse = useFetch<UserPrefPayload, undefined>(getPreference);
 
-	const [firstName, setFirstName] = useState<string>('');
-	const [email, setEmail] = useState<string>('');
-	const [organizationName, setOrganisationName] = useState<string>('');
-	const [keepMeUpdated, setKeepMeUpdated] = useState<boolean>(true);
-	const [anonymise, setAnonymise] = useState<boolean>(false);
+	if (versionResponse.error || userPrefResponse.error) {
+		return (
+			<Typography>
+				{versionResponse.errorMessage ||
+					userPrefResponse.errorMessage ||
+					'Somehthing went wrong'}
+			</Typography>
+		);
+	}
 
-	const setState = (
-		value: string,
-		setFunction: React.Dispatch<React.SetStateAction<string>>,
-	) => {
-		setFunction(value);
-	};
+	if (
+		versionResponse.loading ||
+		versionResponse.payload === undefined ||
+		userPrefResponse.loading ||
+		userPrefResponse.payload === undefined
+	) {
+		return <Spinner tip="Loading.." />;
+	}
 
-	const handleSubmit = (e: React.FormEvent<HTMLFormElement>): void => {
-		(async (): Promise<void> => {
-			try {
-				e.preventDefault();
-				setLoading(true);
-				const payload = {
-					first_name: firstName,
-					email: email,
-				};
+	const version = versionResponse.payload.version;
 
-				const response = await signup({
-					email: payload.email,
-					name: payload.first_name,
-				});
+	const userpref = userPrefResponse.payload;
 
-				if (response.statusCode === 200) {
-					loggedIn();
-					history.push(ROUTES.APPLICATION);
-				} else {
-					notificationsInstance.error({
-						message: 'Something went wrong',
-					});
-				}
-				setLoading(false);
-			} catch (error) {
-				notificationsInstance.error({
-					message: 'Something went wrong',
-				});
-				setLoading(false);
-			}
-		})();
-	};
-
-	const onSwitchHandler = (
-		value: boolean,
-		setFunction: React.Dispatch<React.SetStateAction<boolean>>,
-	) => {
-		setFunction(value);
-	};
-
-	return (
-		<Container>
-			{Element}
-
-			<LeftContainer direction="vertical">
-				<Space align="center">
-					<Logo src={'signoz.svg'} alt="logo" />
-					<Title style={{ fontSize: '46px', margin: 0 }}>SigNoz</Title>
-				</Space>
-				<Typography>
-					Monitor your applications. Find what is causing issues.
-				</Typography>
-				<Card
-					style={{ width: 'max-content' }}
-					bodyStyle={{ padding: '1px 8px', width: '100%' }}
-				>
-					SigNoz v0.5.4
-				</Card>
-			</LeftContainer>
-
-			<FormWrapper>
-				<form onSubmit={handleSubmit}>
-					<Title level={4}>Create your account</Title>
-					<div>
-						<Label htmlFor="signupEmail">Email</Label>
-						<Input
-							placeholder="mike@netflix.com"
-							type="email"
-							autoFocus
-							value={email}
-							onChange={(e): void => {
-								setState(e.target.value, setEmail);
-							}}
-							required
-							id="signupEmail"
-						/>
-					</div>
-
-					<div>
-						<Label htmlFor="signupFirstName">First Name</Label>
-						<Input
-							placeholder="Mike"
-							value={firstName}
-							onChange={(e): void => {
-								setState(e.target.value, setFirstName);
-							}}
-							required
-							id="signupFirstName"
-						/>
-					</div>
-					<div>
-						<Label htmlFor="organizationName">Organization Name</Label>
-						<Input
-							placeholder="Netflix"
-							value={organizationName}
-							onChange={(e): void => {
-								setState(e.target.value, setOrganisationName);
-							}}
-							required
-							id="organizationName"
-						/>
-					</div>
-
-					<MarginTop marginTop={'2.4375rem'}>
-						<Space>
-							<Switch
-								onChange={(value) => onSwitchHandler(value, setKeepMeUpdated)}
-								checked={keepMeUpdated}
-							/>
-							<Typography>Keep me updated on new SigNoz features</Typography>
-						</Space>
-					</MarginTop>
-
-					<MarginTop marginTop={'0.5rem'}>
-						<Space>
-							<Switch
-								onChange={(value) => onSwitchHandler(value, setAnonymise)}
-								checked={anonymise}
-							/>
-							<Typography>
-								Anonymise my usage date. We collect data to measure product usage
-							</Typography>
-						</Space>
-					</MarginTop>
-
-					<ButtonContainer>
-						<Button
-							type="primary"
-							htmlType="submit"
-							data-attr="signup"
-							loading={loading}
-							disabled={loading || !email || !organizationName || !firstName}
-						>
-							Get Started
-						</Button>
-					</ButtonContainer>
-				</form>
-			</FormWrapper>
-		</Container>
-	);
+	return <SignUpComponent userpref={userpref} version={version} />;
 };
 
-interface DispatchProps {
-	loggedIn: () => void;
-}
-
-const mapDispatchToProps = (
-	dispatch: ThunkDispatch<unknown, unknown, AppActions>,
-): DispatchProps => ({
-	loggedIn: bindActionCreators(UserLoggedIn, dispatch),
-});
-
-type SignupProps = DispatchProps;
-
-export default connect(null, mapDispatchToProps)(Signup);
+export default SignUp;
