+++ frontend/src/store/actions/app/index.ts
@@ -1,3 +1,2 @@
 export * from './toggleDarkMode';
 export * from './toggleSettingsTab';
-export * from './userLoggedIn';
