+++ frontend/src/types/api/user/getVersion.ts
@@ -1,0 +1,3 @@
+export interface PayloadProps {
+	version: string;
+}
